Change FlightId type to int

+++ src/main/java/za/ca/cput/adp3capstone/domain/Flight.java
@@ -1,7 +1,7 @@
 package za.ca.cput.adp3capstone.domain;
 
 public class Flight {
-    private double flightId;
+    private int flightId;
     private String departure_airport;
     private String arrival_airport;
     private int duration;
@@ -13,7 +13,7 @@ public class Flight {
     private boolean overnight;
     private double price;
 
-    public double getFlightId() {
+    public int getFlightId() {
         return flightId;
     }
 
@@ -72,7 +72,7 @@ public class Flight {
     }
 
     public static class Builder{
-        private double flightId;
+        private int flightId;
         private String departure_airport;
         private String arrival_airport;
         private int duration;
@@ -84,7 +84,7 @@ public class Flight {
         private boolean overnight;
         private double price;
 
-        public Builder setFlightId(double  flightId) {
+        public Builder setFlightId(int flightId) {
             this.flightId = flightId;
             return this;
         }
