+++ src/main/java/za/ca/cput/adp3capstone/repisitory/IFlightRepository.java
@@ -4,6 +4,6 @@ import za.ca.cput.adp3capstone.domain.Flight;
 
 import java.util.HashMap;
 
-public interface IFlightRepository extends IRepository<Flight, Double> {
-    public HashMap<Double, Flight> getAll();
+public interface IFlightRepository extends IRepository<Flight, Integer> {
+    public HashMap<Integer, Flight> getAll();
 }
